fix: mongo repository component scan
