+++ pom.xml
@@ -18,7 +18,6 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<java.version>1.8</java.version>
-		<keycloak.version>3.1.0.Final</keycloak.version>
 	</properties>
 
 	<dependencies>
