+++ src/main/java/com/tinderbot/entities/requests/SignUpRequest.java
@@ -1,12 +1,15 @@
-package com.tinderbot.entities;
+package com.tinderbot.entities.requests;
 
-import javax.validation.constraints.Size;
+import javax.validation.constraints.*;
 
-public class LoginForm {
-    
-	@Size(min=3, max = 60)
+public class SignUpRequest {
+
+    @Size(min = 3, max = 50)
     private String username;
 
+    @Size(max = 60)
+    private String email;
+       
     @Size(min = 6, max = 40)
     private String password;
 
@@ -18,6 +21,14 @@ public class LoginForm {
         this.username = username;
     }
 
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
     public String getPassword() {
         return password;
     }
@@ -25,4 +36,5 @@ public class LoginForm {
     public void setPassword(String password) {
         this.password = password;
     }
+
 }
