+++ src/main/resources/application.properties
@@ -12,6 +12,8 @@ tinder.url.superLikeUser=https://api.gotinder.com/like/
 tinder.url.updates=https://api.gotinder.com/updates
 tinder.url.userProfile=https://api.gotinder.com/profile
 
+spring.data.mongodb.database=tinderbot
+
 watson.username=
 watson.password=
 watson.workspace-id=
